Implemented Level headerfile & Graph data structure
Implemented a level headerfile to show how the levels will be interfaced
with. A graph data structure has been created to solve how rooms will
know which other rooms are adjacent to each other.

TODO: test it all ffs
TODO: other level-designer only functions
TODO: whether to store graph in the .lvl file or construct on-the-fly