Commit current state,
Further refactored system folder. Found fatal error with queue.c, where memory
overwrites weirdly. To be fixed by implementing the queue as a linked list
instead.

Text should allegedly be able to be drawn to the screen, however this has not
been tested on account of the drawqueue being fucked.